Working on creating the Champion table.

+++ RiotApi/src/main/java/com/howbig/riot/type/champion/Stats.java
@@ -4,24 +4,24 @@ package com.howbig.riot.type.champion;
  * Created by Alex on 5/18/2014.
  */
 public class Stats {
-    public int hp;
-    public int hpperlevel;
-    public int mp;
-    public int mpperlevel;
-    public int movespeed;
-    public int armor;
-    public int armorperlevel;
-    public int spellblock;
-    public int spellblockperlevel;
-    public int attackrange;
-    public int hpregen;
-    public int hpregenperlevel;
-    public int mpregen;
-    public int mpregenperlevel;
-    public int crit;
-    public int critperlevel;
-    public int attackdamage;
-    public int attackdamageperlevel;
-    public int attackspeedoffset;
-    public int attackspeedperlevel;
+    public double hp;
+    public double hpperlevel;
+    public double mp;
+    public double mpperlevel;
+    public double movespeed;
+    public double armor;
+    public double armorperlevel;
+    public double spellblock;
+    public double spellblockperlevel;
+    public double attackrange;
+    public double hpregen;
+    public double hpregenperlevel;
+    public double mpregen;
+    public double mpregenperlevel;
+    public double crit;
+    public double critperlevel;
+    public double attackdamage;
+    public double attackdamageperlevel;
+    public double attackspeedoffset;
+    public double attackspeedperlevel;
 }
